+++ lib/src/file_picker.dart
@@ -65,29 +65,33 @@ abstract class FilePicker extends PlatformInterface {
 
   /// Retrieves the file(s) from the underlying platform
   ///
-  /// Default `type` set to [FileType.any] with `allowMultiple` set to `false`.
-  /// Optionally, `allowedExtensions` might be provided (e.g. `[pdf, svg, jpg]`.).
+  /// Default [type] set to [FileType.any] with [allowMultiple] set to `false`.
+  /// Optionally, [allowedExtensions] might be provided (e.g. `[pdf, svg, jpg]`.).
   ///
-  /// If `withData` is set, picked files will have its byte data immediately available on memory as `Uint8List`
+  /// If [withData] is set, picked files will have its byte data immediately available on memory as `Uint8List`
   /// which can be useful if you are picking it for server upload or similar. However, have in mind that
   /// enabling this on IO (iOS & Android) may result in out of memory issues if you allow multiple picks or
-  /// pick huge files. Use `withReadStream` instead. Defaults to `true` on web, `false` otherwise.
+  /// pick huge files. Use [withReadStream] instead. Defaults to `true` on web, `false` otherwise.
   ///
-  /// If `withReadStream` is set, picked files will have its byte data available as a `Stream<List<int>>`
+  /// If [withReadStream] is set, picked files will have its byte data available as a [Stream<List<int>>]
   /// which can be useful for uploading and processing large files. Defaults to `false`.
   ///
   /// If you want to track picking status, for example, because some files may take some time to be
   /// cached (particularly those picked from cloud providers), you may want to set [onFileLoading] handler
   /// that will give you the current status of picking.
   ///
-  /// If `allowCompression` is set, it will allow media to apply the default OS compression.
+  /// If [allowCompression] is set, it will allow media to apply the default OS compression.
   /// Defaults to `true`.
   ///
-  /// `dialogTitle` can be optionally set on desktop platforms to set the modal window title. It will be ignored on
+  /// If [lockParentWindow] is set, the child window (file picker window) will
+  /// stay in front of the Flutter window until it is closed (like a modal
+  /// window). This parameter works only on Windows desktop.
+  ///
+  /// [dialogTitle] can be optionally set on desktop platforms to set the modal window title. It will be ignored on
   /// other platforms.
   ///
-  /// The result is wrapped in a `FilePickerResult` which contains helper getters
-  /// with useful information regarding the picked `List<PlatformFile>`.
+  /// The result is wrapped in a [FilePickerResult] which contains helper getters
+  /// with useful information regarding the picked [List<PlatformFile>].
   ///
   /// For more information, check the [API documentation](https://github.com/miguelpruivo/flutter_file_picker/wiki/api).
   ///
@@ -101,6 +105,7 @@ abstract class FilePicker extends PlatformInterface {
     bool allowMultiple = false,
     bool withData = false,
     bool withReadStream = false,
+    bool lockParentWindow = false,
   }) async =>
       throw UnimplementedError('pickFiles() has not been implemented.');
 
@@ -119,12 +124,17 @@ abstract class FilePicker extends PlatformInterface {
   /// Selects a directory and returns its absolute path.
   ///
   /// On Android, this requires to be running on SDK 21 or above, else won't work.
-  /// Returns `null` if folder path couldn't be resolved.
-  ///
-  /// `dialogTitle` can be set to display a custom title on desktop platforms. It will be ignored on Web & IO.
-  ///
   /// Note: Some Android paths are protected, hence can't be accessed and will return `/` instead.
-  Future<String?> getDirectoryPath({String? dialogTitle}) async =>
+  ///
+  /// [dialogTitle] can be set to display a custom title on desktop platforms. It will be ignored on Web & IO.
+  ///
+  /// If [lockParentWindow] is set, the child window (file picker window) will
+  /// stay in front of the Flutter window until it is closed (like a modal
+  /// window). This parameter works only on Windows desktop.
+  ///
+  /// Returns `null` if aborted or if the folder path couldn't be resolved.
+  Future<String?> getDirectoryPath(
+          {String? dialogTitle, bool lockParentWindow = false}) async =>
       throw UnimplementedError('getDirectoryPath() has not been implemented.');
 
   /// Opens a save file dialog which lets the user select a file path and a file
@@ -138,20 +148,27 @@ abstract class FilePicker extends PlatformInterface {
   /// Windows).
   ///
   /// [dialogTitle] can be set to display a custom title on desktop platforms.
+  ///
   /// [fileName] can be set to a non-empty string to provide a default file
   /// name.
+  ///
   /// The file type filter [type] defaults to [FileType.any]. Optionally,
   /// [allowedExtensions] might be provided (e.g. `[pdf, svg, jpg]`.). Both
   /// parameters are just a proposal to the user as the save file dialog does
   /// not enforce these restrictions.
   ///
-  /// Returns [null] if aborted. Returns a [Future<String?>] which resolves to
+  /// If [lockParentWindow] is set, the child window (file picker window) will
+  /// stay in front of the Flutter window until it is closed (like a modal
+  /// window). This parameter works only on Windows desktop.
+  ///
+  /// Returns `null` if aborted. Returns a [Future<String?>] which resolves to
   /// the absolute path of the selected file, if the user selected a file.
   Future<String?> saveFile({
     String? dialogTitle,
     String? fileName,
     FileType type = FileType.any,
     List<String>? allowedExtensions,
+    bool lockParentWindow = false,
   }) async =>
       throw UnimplementedError('saveFile() has not been implemented.');
 }