+++ lib/src/file_picker_io.dart
@@ -31,6 +31,7 @@ class FilePickerIO extends FilePicker {
     bool allowMultiple = false,
     bool? withData = false,
     bool? withReadStream = false,
+    bool lockParentWindow = false,
   }) =>
       _getPath(
         type,
@@ -47,7 +48,8 @@ class FilePickerIO extends FilePicker {
       _channel.invokeMethod<bool>('clear');
 
   @override
-  Future<String?> getDirectoryPath({String? dialogTitle}) async {
+  Future<String?> getDirectoryPath(
+      {String? dialogTitle, bool lockParentWindow = false}) async {
     try {
       return await _channel.invokeMethod('dir', {});
     } on PlatformException catch (ex) {